0.3.1 - first attempt with second can bus
- Added inside temperature. C-can: 654h byte 7.
- First data from M-bus.
- Splitted source into five new files.
